Move common storage to root level
