+++ src/app/api/api.ts
@@ -1,7 +1,6 @@
 import axios from 'axios';
 import Søknad from '../types/søknad/Søknad';
 import Environment from '../../app/Environment';
-import { Attachment } from '../types/Attachment';
 
 function getPerson() {
     const endpoint = Environment.REST_API_URL;
@@ -32,29 +31,6 @@ function sendSøknad(søknad: Søknad) {
     });
 }
 
-function saveAttachment(attachment: Attachment) {
-    const config = {
-        withCredentials: true,
-        headers: {
-            'content-type': 'multipart/form-data'
-        }
-    };
-
-    const formData = new FormData();
-    formData.append('vedlegg', attachment.file);
-
-    const url = `${Environment.REST_API_URL}/storage/vedlegg`;
-    return axios.post(url, formData, config);
-}
-
-function deleteAttachment(attachment: Attachment) {
-    const config = {
-        withCredentials: true
-    };
-    const url = `${Environment.REST_API_URL}/storage/vedlegg/${attachment.url}`;
-    return axios.delete(url, config);
-}
-
-const Api = { getPerson, sendSøknad, saveAttachment, deleteAttachment };
+const Api = { getPerson, sendSøknad };
 
 export default Api;
