+++ src/app/components/vedlegg/VedleggListe.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import Vedlegg from './Vedlegg';
-import { Attachment } from '../../types/Attachment';
 import { TransitionGroup, CSSTransition } from 'react-transition-group';
 import './vedlegg.less';
+import { Attachment } from 'storage/attachment/types/Attachment';
 
 interface Props {
     vedlegg: Attachment[];
