+++ src/app/components/vedlegg/VedleggOversikt.tsx
@@ -5,11 +5,11 @@ import VedleggInput from '../vedlegg/VedleggInput';
 import VedleggListe from '../vedlegg/VedleggListe';
 import LabelText from '../labeltekst/Labeltekst';
 import { bytesString, getTotalFileSize } from '../../util/attachment';
-import { Attachment } from '../../types/Attachment';
 import { mapFileToAttachment } from './util';
 import { SøknadsvedleggKey } from '../../types/søknad/Søknadsvedlegg';
 import { CSSTransition } from 'react-transition-group';
 import { guid } from 'nav-frontend-js-utils';
+import { Attachment } from 'storage/attachment/types/Attachment';
 
 export interface VedleggOversiktProps {
     vedlegg: Attachment[];
