+++ src/app/connected-components/steg/relasjon-til-barn-stebarnsadopsjon/RelasjonTilBarnStebarnsadopsjon.tsx
@@ -17,16 +17,15 @@ import Labeltekst from '../../../components/labeltekst/Labeltekst';
 
 import utils from '../../../util/fødselsdato';
 import VedleggOversikt from '../../../components/vedlegg/VedleggOversikt';
-import ISøknadsvedlegg from '../../../types/søknad/Søknadsvedlegg';
+import Søknadsvedlegg from '../../../types/søknad/Søknadsvedlegg';
 import {
     concatNewFiles,
     removeFileFromArray
 } from '../../../components/vedlegg/util';
-import { Attachment } from '../../../types/Attachment';
 
 export interface StateProps {
     barn: Adopsjonsbarn;
-    vedlegg: ISøknadsvedlegg;
+    vedlegg: Søknadsvedlegg;
 }
 
 export type Props = DispatchProps & StateProps & InjectedIntlProps;
@@ -81,7 +80,7 @@ class RelasjonTilBarnStebarnsadopsjon extends React.Component<Props, {}> {
                         <VedleggOversikt
                             inputId="adopsjonsvedtak"
                             vedlegg={vedlegg.adopsjonsvedtak}
-                            onFilesSelect={(files: Attachment[]) => {
+                            onFilesSelect={(files) => {
                                 dispatch(
                                     søknadActions.updateVedlegg({
                                         adopsjonsvedtak: concatNewFiles(
@@ -91,7 +90,7 @@ class RelasjonTilBarnStebarnsadopsjon extends React.Component<Props, {}> {
                                     })
                                 );
                             }}
-                            onFileDelete={(file: Attachment) =>
+                            onFileDelete={(file) =>
                                 dispatch(
                                     søknadActions.updateVedlegg({
                                         adopsjonsvedtak: removeFileFromArray(
