+++ src/app/types/søknad/Søknad.ts
@@ -1,4 +1,4 @@
-import ISøknadsvedlegg from './Søknadsvedlegg';
+import Søknadsvedlegg from './Søknadsvedlegg';
 import AnnenForelder, { AnnenForelderPartial } from './AnnenForelder';
 
 import Utenlandsopphold, { UtenlandsoppholdPartial } from './Utenlandsopphold';
@@ -33,7 +33,7 @@ interface Søknad {
     barn: Barn;
     utenlandsopphold: Utenlandsopphold;
     uttaksplan: Periode[];
-    vedlegg: ISøknadsvedlegg;
+    vedlegg: Søknadsvedlegg;
     søkerRolle: SøkerRolle;
     erSelvstendigNæringsdrivende: boolean;
     erFrilanser: boolean;
@@ -48,7 +48,7 @@ export interface SøknadPartial {
     barn: BarnPartial;
     utenlandsopphold: UtenlandsoppholdPartial;
     uttaksplan?: Periode[];
-    vedlegg: ISøknadsvedlegg;
+    vedlegg: Søknadsvedlegg;
     søkerRolle?: SøkerRolle;
     erSelvstendigNæringsdrivende?: boolean;
     erFrilanser?: boolean;
