+++ src/storage/attachment/saga/attachmentSaga.ts
@@ -1,23 +1,23 @@
-import Api from '../../api/api';
 import { all, call, put, takeEvery } from 'redux-saga/effects';
 import {
     UploadAttachment,
-    AttachmentActionKeys,
-    DeleteAttachment
-} from '../actions/attachment/attachmentActionDefinitions';
+    DeleteAttachment,
+    AttachmentActionKeys
+} from '../redux/attachmentActionDefinitions';
 import {
+    attachmentPending,
     attachmentUploadSuccess,
     attachmentUploadFailed,
-    attachmentPending,
     attachmentDeleteSuccess,
     attachmentDeleteFailed
-} from '../actions/attachment/attachmentActionCreators';
+} from '../redux/attachmentActionCreators';
+import AttachmentApi from '../../api/attachmentApi';
 
 function* saveAttachment(action: UploadAttachment) {
     const { attachment } = action;
     try {
         yield put(attachmentPending(attachment));
-        const response = yield call(Api.saveAttachment, attachment);
+        const response = yield call(AttachmentApi.saveAttachment, attachment);
         const uri: string = response.headers.location || 'mockurl';
         yield put(attachmentUploadSuccess(uri, attachment));
     } catch (error) {
@@ -29,14 +29,14 @@ function* deleteAttachment(action: DeleteAttachment) {
     const { attachment } = action;
     try {
         yield put(attachmentPending(attachment));
-        yield call(Api.deleteAttachment, attachment);
+        yield call(AttachmentApi.deleteAttachment, attachment);
         yield put(attachmentDeleteSuccess(attachment));
     } catch (error) {
         yield put(attachmentDeleteFailed(error, attachment));
     }
 }
 
-export default function* sagas() {
+export default function* attachmentSaga() {
     yield all([takeEvery(AttachmentActionKeys.UPLOAD, saveAttachment)]);
     yield all([takeEvery(AttachmentActionKeys.DELETE, deleteAttachment)]);
 }
