+++ tsconfig.json
@@ -20,7 +20,8 @@
         "baseUrl": "./src",
         "typeRoots": ["./node_modules/@types", "./src/@types"],
         "paths": {
-            "uttaksplan/*": ["./uttaksplan/*"]
+            "uttaksplan/*": ["./uttaksplan/*"],
+            "storage/*": ["./storage/*"]
         }
     },
     "awesomeTypescriptLoaderOptions": {
